requirements.txt
certifi==2019.3.9
chardet==3.0.4
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==19.9.0
idna==2.8
Pillow==6.0.0
psycopg2==2.7.5
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
urllib3==1.25.2
whitenoise==4.1.2
